# performance (XLStyle): Custom GetHashCode to improve performance on XLStyle structs