## Summary
Custom GetHashCode to improve performance on XLStyle structs  Slight reduction in memory allocation.  vs using `HashCode()`


## Benchmarks

**NOTE**: These benchmarks are from when I stacked this PR ontop of #2674 and execute the benchmark test in that PR, but it does illustrate the change.  

#### Before

| Method          | Mean     | Error    | StdDev   | Gen0      | Gen1     | Allocated |
|---------------- |---------:|---------:|---------:|----------:|---------:|----------:|
| CellFillAndCopy | 43.96 ms | 0.875 ms | 0.899 ms | 1700.0000 | 800.0000 |  28.17 MB |

#### After GetHashCode Changes.

| Method          | Mean     | Error    | StdDev   | Gen0      | Gen1     | Allocated |
|---------------- |---------:|---------:|---------:|----------:|---------:|----------:|
| CellFillAndCopy | 41.08 ms | 0.818 ms | 1.475 ms | 1692.3077 | 615.3846 |  27.99 MB |



## Why HashCode may be slower
It has extra state and indirection
HashCode is a mutable struct that maintains internal buffers
Each .`Add()` writes into the buffer
When calling `.ToHashCode()` → it combines state + random salt → extra CPU cost